carve out mockSnowflakeServiceClient for other testing
